split properties out by spec version
