+++ stix2/v21/bundle.py
@@ -2,7 +2,7 @@ from collections import OrderedDict
 
 from stix2 import parse
 from stix2.base import _STIXBase
-from stix2.properties import IDProperty, ListProperty, Property, TypeProperty
+from .properties import IDProperty, ListProperty, Property, TypeProperty
 from stix2.utils import _get_dict, get_class_hierarchy_names
