+++ stix2/v21/sro.py
@@ -4,9 +4,9 @@ from collections import OrderedDict
 
 from ..base import _STIXBase
 from ..markings import _MarkingsMixin
-from ..properties import (BooleanProperty, IDProperty, IntegerProperty,
-                          ListProperty, ReferenceProperty, StringProperty,
-                          TimestampProperty, TypeProperty)
+from .properties import (BooleanProperty, IDProperty, IntegerProperty,
+                         ListProperty, ReferenceProperty, StringProperty,
+                         TimestampProperty, TypeProperty)
 from ..utils import NOW
 from .common import ExternalReference, GranularMarking
